[Silver III] Title: 주유소, Time: 84 ms, Memory: 3584 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -3,7 +3,7 @@
 using namespace std;
 
 int n;
-int dist[100002], cost[100002];
+long long dist[100002], cost[100002];
 int main() {
 	cin >> n;
 	for (int i = 0; i < n-1; i++) {
@@ -13,8 +13,8 @@ int main() {
 		cin >> cost[i];
 	}
 	long long ans = 0;
-	int tempCost = cost[0];
-	int tempDist = dist[0];
+	long long tempCost = cost[0];
+	long long tempDist = dist[0];
 	for (int i = 1; i < n - 1; i++) {
 		if (tempCost > cost[i]) {
 			ans += tempCost * tempDist;
